pyproject.toml
[tool.poetry]
name = "minizinc-discord"
version = "0.1.0"
description = "A Discord bot that will solve MiniZinc instances"
authors = ["Jip J. Dekker <jip@dekker.one>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^2.0.1"
minizinc = "^0.7.0"

[tool.poetry.dev-dependencies]
black = "^22.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
